requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
aiosqlite==0.21.0
beautifulsoup4==4.14.3
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
Requests==2.32.5
Unidecode==1.4.0
yookassa==3.6.0
